Separate library code from userland code
