+++ src/web/backend/layers/templated/templated.ts
@@ -1,4 +1,4 @@
-import {alphabet} from "../../alphabet";
+import {alphabet} from "/lib/alphabet/alphabet";
 import {underlying} from "../underlying/underlying";
 
 /*
