+++ src/web/frontend/assets/global.css
