+++ tsconfig.json
@@ -22,5 +22,11 @@
         "esModuleInterop": true,
         "skipLibCheck": true,
         "forceConsistentCasingInFileNames": true,
+        "baseUrl": "src",
+        "paths": {
+            "/*": [
+                "./*"
+            ]
+        }
     }
 }
